+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## Angular Gamification Changelog
 
+<a name="Fev 16, 2018"></a>
+### Fev 16, 2018
+* Accepting module config
+
 <a name="Fev 14, 2018"></a>
 ### Fev 14, 2018
 * Create library